# docs/api/scheduling.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EdTech tutoring API",
    "version": "1.0.0",
    "description": "Контракт регистрации, JWT-аутентификации, профилей, приглашений и расписания. Access JWT передаётся как Bearer token, refresh token — только в HttpOnly cookie."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/auth/email-verification/confirm": {
      "post": {
        "operationId": "confirmEmail",
        "tags": [
          "Auth"
        ],
        "summary": "Подтвердить email",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email подтверждён. Выпущен access JWT и установлена refresh cookie.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "Устанавливает или ротирует refresh token в HttpOnly cookie REFRESH_TOKEN. В production cookie имеет Secure; атрибуты SameSite и Path задаются сервером.",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "description": "Ответ с токеном не должен сохраняться HTTP-кэшами.",
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Pragma": {
                "description": "Совместимый запрет кэширования ответа с токеном.",
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-cache"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Ошибка формата или валидации.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/email-verification/resend": {
      "post": {
        "operationId": "resendEmailVerification",
        "tags": [
          "Auth"
        ],
        "summary": "Повторно отправить подтверждение",
        "description": "Для неподтверждённого CURRENT email создаёт новую REGISTRATION verification, для ACTIVE account с совпавшим PENDING email — EMAIL_CHANGE. Предыдущая активная verification того же purpose инвалидируется. Ответ не раскрывает существование или состояние аккаунта.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendEmailVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Запрос принят независимо от существования аккаунта."
          },
          "400": {
            "description": "Ошибка формата или валидации.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/me": {
      "get": {
        "operationId": "getMe",
        "tags": [
          "Profiles"
        ],
        "summary": "Получить текущего пользователя и профили",
        "responses": {
          "200": {
            "description": "Текущий пользователь.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не аутентифицирован.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateMe",
        "tags": [
          "Profiles"
        ],
        "summary": "Изменить общие данные текущего пользователя",
        "description": "При смене email текущий адрес сохраняется в email, новый записывается в pendingEmail, а письмо EMAIL_CHANGE ставится в очередь до фиксации той же транзакции. Повторное письмо запрашивается отдельным resend endpoint.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Данные обновлены.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка формата или валидации.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Не аутентифицирован.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Действие запрещено.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Короткоживущий access JWT из TokenResponse."
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "REFRESH_TOKEN",
        "description": "Refresh token в HttpOnly cookie. Используется только endpoint-ами refresh и logout."
      }
    },
...
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "TEACHER",
          "STUDENT"
        ]
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "PENDING_EMAIL_VERIFICATION",
          "ACTIVE",
          "SUSPENDED",
          "DEACTIVATED"
        ]
      },
      "UserResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Непрозрачный идентификатор User."
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "pendingEmail": {
            "type": "string",
            "format": "email",
            "maxLength": 254,
            "description": "Новый email, ожидающий подтверждения. Текущий email остаётся в поле email до подтверждения.",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Непустое после trim."
          },
          "lastName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Непустое после trim."
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRole"
            },
            "minItems": 1,
            "maxItems": 2,
            "uniqueItems": true
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "emailVerifiedAt": {
            "type": "string",
            "format": "date-time",
            "description": "RFC 3339 с обязательным Z или UTC-offset; сервер возвращает UTC с Z.",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Момент создания аккаунта; сервер возвращает UTC с Z."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Момент последнего изменения аккаунта; сервер возвращает UTC с Z."
          }
        },
        "required": [
          "id",
          "email",
          "pendingEmail",
          "firstName",
          "lastName",
          "roles",
          "status",
          "emailVerifiedAt",
          "createdAt",
          "updatedAt"
        ],
        "example": {
          "id": "u_002",
          "email": "anna@example.com",
          "pendingEmail": null,
          "firstName": "Анна",
          "lastName": "Петрова",
          "roles": [
            "STUDENT"
          ],
          "status": "ACTIVE",
          "emailVerifiedAt": "2026-09-06T12:00:00Z",
          "createdAt": "2026-09-06T11:55:00Z",
          "updatedAt": "2026-09-06T12:00:00Z"
        }
      },
...
      "ConfirmEmailRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string",
            "minLength": 20,
            "maxLength": 512,
            "writeOnly": true
          }
        },
        "required": [
          "token"
        ],
        "example": {
          "token": "vfy_12345678901234567890"
        }
      },
      "ResendEmailVerificationRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          }
        },
        "required": [
          "email"
        ],
        "example": {
          "email": "anna@example.com"
        }
      },
...
      "TokenResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "accessToken": {
            "type": "string",
            "minLength": 20,
            "maxLength": 8192,
            "description": "Короткоживущий JWT access token. Frontend передаёт его как Authorization: Bearer <token>."
          },
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expiresInSeconds": {
            "type": "integer",
            "minimum": 1,
            "description": "Число секунд до окончания срока действия access token."
          }
        },
        "required": [
          "accessToken",
          "tokenType",
          "expiresInSeconds"
        ],
        "description": "Refresh token не включается в JSON и передаётся только через HttpOnly cookie.",
        "example": {
          "accessToken": "eyJhbGciOiJSUzI1NiJ9.eyJzdWIiOiJ1XzAwMiJ9.example-signature",
          "tokenType": "Bearer",
          "expiresInSeconds": 900
        }
      },
      "UpdateMeRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Непустое после trim."
          },
          "lastName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Непустое после trim."
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          }
        },
        "minProperties": 1,
        "description": "Минимум одно поле. Роли и пароль не изменяются.",
        "example": {
          "firstName": "Анна-Мария",
          "email": "anna.new@example.com"
        }
      },
...
      "TeacherProfileResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "subjectCodes": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 50,
              "pattern": "^[A-Z][A-Z0-9_]*$"
            },
            "minItems": 1,
            "uniqueItems": true
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "education": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "experienceYears": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "nullable": true
          },
          "city": {
            "type": "string",
            "maxLength": 100,
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "format": "uri",
            "maxLength": 2048,
            "nullable": true
          }
        },
        "required": [
          "user",
          "subjectCodes",
          "description",
          "education",
          "experienceYears",
          "city",
          "photoUrl"
        ]
      },
...
      "StudentProfileResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "subjectCodes": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 50,
              "pattern": "^[A-Z][A-Z0-9_]*$"
            },
            "minItems": 1,
            "uniqueItems": true
          },
          "photoUrl": {
            "type": "string",
            "format": "uri",
            "maxLength": 2048,
            "nullable": true
          }
        },
        "required": [
          "user",
          "birthDate",
          "subjectCodes",
          "photoUrl"
        ]
      },
      "MeResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "teacherProfile": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TeacherProfileResponse"
              }
            ],
            "nullable": true
          },
          "studentProfile": {
            "allOf": [
              {
                "$ref": "#/components/schemas/StudentProfileResponse"
              }
            ],
            "nullable": true
          }
        },
        "required": [
          "user",
          "teacherProfile",
          "studentProfile"
        ]
      },
...
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ]
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "VALIDATION_ERROR",
          "UNKNOWN_SUBJECT",
          "UNAUTHENTICATED",
          "INVALID_CREDENTIALS",
          "EMAIL_NOT_VERIFIED",
          "INVALID_REFRESH_TOKEN",
          "FORBIDDEN",
          "NOT_FOUND",
          "PROFILE_NOT_FOUND",
          "EMAIL_ALREADY_EXISTS",
          "INVALID_VERIFICATION_TOKEN",
          "INVITATION_ALREADY_PENDING",
          "INVITATION_EXPIRED",
          "INVALID_INVITATION_STATE",
          "ALREADY_LINKED",
          "STUDENT_NOT_LINKED",
          "LESSON_OVERLAP",
          "INVALID_LESSON_STATE",
          "RATE_LIMIT_EXCEEDED",
          "INTERNAL_ERROR"
        ]
      },
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "requestId": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "code",
          "message",
          "fieldErrors",
          "requestId"
        ],
        "example": {
          "code": "LESSON_OVERLAP",
          "message": "Участник уже занят в это время",
          "fieldErrors": [],
          "requestId": "req_001"
        }
      }
    }
  }
}